Resolve the Comments Mentioned
