+++ src/App.js
@@ -1,65 +1,12 @@
-// import React from 'react';
-// import { BrowserRouter, Routes, Route } from 'react-router-dom';
-// import './App.scss';
-// import Navbar from './components/Navbar';
-// import TabsComponent from './components/TabsComponent';
-// import PatientList from './components/PatientList';
-// import PatientDashboard from './Patient_Dashboard/PatientDashboard';
-// import UploadDocumentPage from './Patient_Dashboard/UploadDocumentPage';
-// import SearchButton from './components/SearchButton';
-
-// function App() {
-//   return (
-//     <BrowserRouter>
-//       <div className="app-container">
-//         <Navbar />
-//         <div className="content">
-//           <Routes>
-//             <Route path="/" element={<HomeLayout />} />
-//             <Route path="/bahmni/*" element={<BahmniRedirect />} />
-//             <Route path="/patient_dashboard/:id" element={<PatientDashboard />} />
-//             <Route path="/upload/:visitId" element={<UploadDocumentPage />} />
-//           </Routes>
-//         </div>
-//       </div>
-//     </BrowserRouter>
-//   );
-// }
-
-// const HomeLayout = () => {
-//   return (
-//     <>
-//       <div className="tabs-container">
-//         <TabsComponent />
-//       </div>
-//       <div className="search-container">
-//         <SearchButton />
-//       </div>
-//       <Routes>
-//         <Route path="/" element={<PatientList />} />
-//       </Routes>
-//     </>
-//   );
-// };
-
-// const BahmniRedirect = () => {
-//   window.location.href = 'https://localhost/bahmni/home/#/dashboard';
-//   return null;
-// };
-
-// export default App;
-
-
-
 import React from 'react';
 import { BrowserRouter, Routes, Route } from 'react-router-dom';
 import './App.scss';
-import Navbar from './components/Navbar';
-import TabsComponent from './components/TabsComponent';
-import PatientList from './components/PatientList';
-import PatientDashboard from './Patient_Dashboard/PatientDashboard';
-import UploadDocumentPage from './Patient_Dashboard/UploadDocumentPage';
-import SearchButton from './components/SearchButton';
+import Navbar from './components/Navbar/Navbar';
+import TabsComponent from './components/TabsComponent/TabsComponent';
+import PatientList from './components/PatientList/PatientList';
+import PatientDashboard from './Patient_Dashboard/PatientDashboard/PatientDashboard';
+import UploadDocumentPage from './Patient_Dashboard/UploadDocumentPage/UploadDocumentPage';
+import SearchButton from './components/SearchButton/SearchButton';
 
 function App() {
   return (
@@ -69,7 +16,6 @@ function App() {
         <div className="content">
           <Routes>
             <Route path="/" element={<HomeLayout />} />
-            <Route path="/bahmni/*" element={<BahmniRedirect />} />
             <Route path="/patient_dashboard/:id" element={<PatientDashboard />} />
             <Route path="/upload/:visitId" element={<UploadDocumentPage />} />
           </Routes>
@@ -94,10 +40,4 @@ const HomeLayout = () => {
     </>
   );
 };
-
-const BahmniRedirect = () => {
-  window.location.href = 'https://localhost/bahmni/home/#/dashboard';
-  return null;
-};
-
 export default App;
