+++ src/Patient_Dashboard/DashboardLayout/DashboardLayout.js
@@ -1,11 +1,10 @@
 import React from 'react';
-import Navbar from '../components/Navbar';
+import Navbar from '../../components/Navbar/Navbar';
 
 const DashboardLayout = ({ children }) => {
   return (
     <>
       <Navbar />
-
       <div className="dashboard-container">
         {children}
       </div>
