+++ src/Patient_Dashboard/PatientDashboard/PatientDashboard.scss
@@ -15,8 +15,8 @@
   .patient-card-container {
     display: flex;
     align-items: center;
-    background-color: rgba(128, 128, 128, 0.167);
-    color: #fff;
+    background-color: rgba(211, 199, 199, 0.167);
+    // color: #fff;
     margin-bottom: 1rem;
     flex: 1;
     padding: 1rem;
