+++ src/Patient_Dashboard/UploadDocumentPage/PatientName.js
@@ -1,0 +1,8 @@
+// PatientName.js
+import React from "react";
+
+const PatientName = ({ patientName }) => {
+  return <div className="patient-name">{patientName}</div>;
+};
+
+export default PatientName;
