+++ src/components/PatientCard/PatientCard.module.scss
